+++ PowerPointLabs/PowerPointLabs/ActionFramework/PictureSlidesLab/PictureSlidesEnabledHandler.cs
@@ -1,9 +1,8 @@
 ﻿using PowerPointLabs.ActionFramework.Common.Attribute;
 using PowerPointLabs.ActionFramework.Common.Interface;
 using PowerPointLabs.TextCollection;
-using PowerPointLabs.TimerLab;
 
-namespace PowerPointLabs.ActionFramework.TimerLab
+namespace PowerPointLabs.ActionFramework.PictureSlidesLab
 {
     [ExportEnabledRibbonId(PictureSlidesLabText.PaneTag)]
     class PictureSlidesEnabledHandler : EnabledHandler
