+++ PowerPointLabs/PowerPointLabs/PictureSlidesLab/PictureSlidesLab.cs
@@ -2,7 +2,6 @@
 {
     internal static class PictureSlidesLab
     {
-#pragma warning disable 0618
         public static bool IsPictureSlidesEnabled { get; set; } = false;
     }
 }
